Refactor wdoc loading into services
